requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.8.1
django-media-pil==1.0.0
Pillow==7.0.0
pkg-resources==0.0.0
pycrop==1.0.1
pytz==2019.3
sqlparse==0.3.0
Unidecode==1.1.1
